#89 [feat] 다이얼로그 패키지 분리
