+++ app/src/main/java/com/gojol/notto/ui/home/dialog/TodoDeletionDialog.kt
@@ -1,4 +1,4 @@
-package com.gojol.notto.ui.home
+package com.gojol.notto.ui.home.dialog
 
 import android.app.Dialog
 import android.content.Context
@@ -8,41 +8,34 @@ import android.view.LayoutInflater
 import androidx.core.content.ContextCompat
 import androidx.databinding.DataBindingUtil
 import com.gojol.notto.R
-import com.gojol.notto.databinding.DialogTodoRepeatTimeBinding
-import java.util.*
+import com.gojol.notto.databinding.DialogDeletionBinding
 
 
-class TodoRepeatTimeDialog(context: Context) : View(context) {
+class TodoDeletionDialog(context: Context) : View(context) {
 
-    private val binding: DialogTodoRepeatTimeBinding = DataBindingUtil.inflate(
-        LayoutInflater.from(context), R.layout.dialog_todo_repeat_time,
+    private val binding: DialogDeletionBinding = DataBindingUtil.inflate(
+        LayoutInflater.from(context), R.layout.dialog_deletion,
         null,
         false
     )
     private var dialog: Dialog = Dialog(context)
 
     init {
-
         dialog.requestWindowFeature(Window.FEATURE_NO_TITLE)
         dialog.setContentView(binding.root)
         dialog.window?.setBackgroundDrawable(ContextCompat.getDrawable(context, R.drawable.bg_dialog))
 
         initClickListener()
-        initCalendar()
         show()
     }
 
     private fun initClickListener() {
-        binding.btnDialogRepeatTimeConfirm.setOnClickListener {
+        binding.btnDialogDeletionConfirm.setOnClickListener {
             confirm()
         }
-        binding.btnDialogRepeatTimeReject.setOnClickListener {
+        binding.btnDialogDeletionReject.setOnClickListener {
             dismiss()
         }
-    }
-
-    private fun initCalendar() {
-        binding.cvRepeatTime.setSelectedDate(Date(System.currentTimeMillis()))
     }
 
     fun show() {
@@ -50,7 +43,7 @@ class TodoRepeatTimeDialog(context: Context) : View(context) {
     }
 
     private fun confirm() {
-        val selectedDate = binding.cvRepeatTime.selectedDate
+
     }
 
     private fun dismiss() {
