+++ app/src/main/res/layout/dialog_todo_repeat_type.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="dialog"
-            type="com.gojol.notto.ui.home.TodoRepeatTypeDialog" />
+            type="com.gojol.notto.ui.home.dialog.TodoRepeatTypeDialog" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
