+++ app/src/main/res/layout/dialog_todo_set_time.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="dialog"
-            type="com.gojol.notto.ui.home.TodoSetTimeDialog" />
+            type="com.gojol.notto.ui.home.dialog.TodoSetTimeDialog" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
